pyproject.toml
[project]
name = "prompster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "py-llmify>=0.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "spotifyify>=0.1.0",
]
